Change windows build to VS2015
Was previously just using mingw since visual studio hadn't been free nor
fully support C++11

Fix a bunch of warnings from VS compiler

Remove windows support from build system

+++ ChessMateEngine/Util/Logging/Log.h
@@ -39,7 +39,7 @@ namespace Util {
  */
 struct Log
 {
-    unsigned int m_index = UINT_MAX;
+    unsigned long long int m_index = UINT_MAX;
     LogLevel m_level = NUM_LEVELS;
     double m_time = -1.0;
     int m_gameId = -1;
